rewrite: turn into a complete Kubernetes operator
The auth service now actively listens to all Stunner configmaps in the Kubernetes cluster and can
render a TURN auth token or an ICE config for each. The generated URIs can be filtered by
namespace, gateway and listener, in which case only the URI for the corresponding namespace,
gateway and listener, respectively, will be generated.